Implement admin commands, and a anti-swear
Access to the admin commands are granted when the client has been opped using 'op <clientName>' which is also a new command. They can also be deopped using another new command 'deop <clientName>', though this command is case-sensitive because of how the chat works. The anti-swear has finally be implemented to keep people clean in the chat.